Adding averageValue implementation at ResultsController

+++ engine/src/engine/Engine.java
@@ -119,11 +119,11 @@ public class Engine implements EngineInterface, Serializable {
         return new DTOSimulationHistogram(values, propertyName);
     }
     @Override
-    public Collection<Integer> getTicksOfSameValueOfPropertyInstances(int id, String propertyName, String entityName) {
-        List<Integer> values = new ArrayList<>();
+    public Collection<Double> getTicksOfSameValueOfPropertyInstances(int id, String propertyName, String entityName) {
+        List<Double> values = new ArrayList<>();
         World world = pastSimulations.get(id).getWorld();
         for (EntityInstance entityInstance : world.getEntityManager().getEntityInstances(world.getEntityManager().getEntityDefinition(entityName))) {
-            int ticksOfSameValue = entityInstance.getPropertyByName(propertyName).getTicksOfSameValue();
+            double ticksOfSameValue = entityInstance.getPropertyByName(propertyName).getTicksOfSameValue();
             values.add(ticksOfSameValue);
         }
         return values;
